Estilos al boton y calendario
